feat: Task 시간대 선호 미지정 상태를 허용하도록 변경, 보정 기능 제거

+++ src/main/java/me/timetablescheduler/domain/preference/PreferenceRepository.java
@@ -2,5 +2,5 @@ package me.timetablescheduler.domain.preference;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PreferenceRepository extends JpaRepository<Preference, Integer> {
+public interface PreferenceRepository extends JpaRepository<Preference, Long> {
 }
